requirements.txt
flask
flask-cors
google-generativeai
google-cloud-secret-manager
python-dotenv
werkzeug